chore: Adding Windows + tflint run
