+++ .github/scripts/setup/windows-setup.ps1
@@ -9,3 +9,8 @@ mkdir C:\bin
 cmd /c mklink C:\bin\sh.exe "C:\Program Files\Git\usr\bin\bash.exe"
 cmd /c mklink C:\bin\bash.exe "C:\Program Files\Git\usr\bin\bash.exe"
 echo "C:\bin" | Out-File -Append -FilePath $env:GITHUB_PATH
+
+# tflint is only needed by the tflint integration tests, so it is installed here
+# rather than in mise.toml to keep local installs lean.
+mise use "tflint@0.50.3"
+tflint --version
