+++ .github/workflows/integration-test.yml
@@ -105,7 +105,7 @@ jobs:
             target: ./...
             setup_scripts:
               - .github/scripts/setup/windows-setup.ps1
-            tags: windows
+            tags: "windows,tflint"
             run: "^TestWindows"
           - name: Provider Cache Server with Latest Terraform
             os: ubuntu
