+++ test/integration_tflint_test.go
@@ -1,5 +1,5 @@
-//go:build tflint
-// +build tflint
+//go:build tflint && !windows
+// +build tflint,!windows
 
 //nolint:paralleltest
 package test_test
