+++ test/integration_windows_test.go
@@ -10,7 +10,6 @@ import (
 	"io/fs"
 	"os"
 	"path/filepath"
-	"regexp"
 	"strings"
 	"testing"
 
@@ -24,7 +23,6 @@ const (
 	testFixtureDownloadPath                         = "fixtures/download"
 	testFixtureLocalRelativeArgsWindowsDownloadPath = "fixtures/download/local-windows"
 	testFixtureManifestRemoval                      = "fixtures/manifest-removal"
-	testFixtureTflintNoIssuesFound                  = "fixtures/tflint/no-issues-found"
 	testFixtureProviderCacheWindowsRemoteURL        = "fixtures/provider-cache/windows-remote-url"
 
 	tempDir = `C:\tmp`
@@ -127,27 +125,6 @@ func TestWindowsTerragruntSourceMapDebug(t *testing.T) {
 			helpers.RunTerragrunt(t, tgArgs)
 		})
 	}
-}
-
-// Get rid of this once we have no internal tflint
-func TestWindowsTflintIsInvoked(t *testing.T) {
-	out := new(bytes.Buffer)
-	errOut := new(bytes.Buffer)
-	rootPath := CopyEnvironmentWithTflint(t, testFixtureTflintNoIssuesFound)
-	modulePath := filepath.Join(rootPath, testFixtureTflintNoIssuesFound)
-	err := helpers.RunTerragruntCommand(t, fmt.Sprintf("terragrunt plan --log-level debug --working-dir %s", modulePath), out, errOut)
-	assert.NoError(t, err)
-
-	assert.NotContains(t, errOut.String(), "Error while running tflint with args:")
-	assert.NotContains(t, errOut.String(), "Tflint found issues in the project. Check for the tflint logs above.")
-
-	// TFLint config should be found in the original working directory, not inside .terragrunt-cache
-	// The config path should end with .tflint.hcl but NOT be inside .terragrunt-cache
-	// Use cross-platform regex patterns that handle both Unix / and Windows \ path separators
-	found, err := regexp.MatchString(`--config\s+[^\s]*\.tflint\.hcl`, errOut.String())
-	assert.NoError(t, err)
-	assert.True(t, found, "Expected tflint to be invoked with --config pointing to .tflint.hcl")
-	assert.NotRegexp(t, `--config\s+[^\s]*[/\\]?\.terragrunt-cache`, errOut.String(), "TFLint config should not be inside cache directory")
 }
 
 func TestWindowsManifestFileIsRemoved(t *testing.T) {
